+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Thinking support. [#47](https://github.com/sinaptia/ruby_llm-evals/pull/47) [@patriciomacadden](https://github.com/patriciomacadden)
+- Pinned runs. [#52](https://github.com/sinaptia/ruby_llm-evals/pull/52) [@patriciomacadden](https://github.com/patriciomacadden)
 
 ### Changed